# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/customers/register": {
      "post": {
        "operationId": "Auth_registerCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthRegistrationAcceptedDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "A customer account cannot be created with the supplied email or phone number.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/customers/login": {
      "post": {
        "operationId": "Auth_loginCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthLoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The login credentials are invalid. Missing, locked, passwordless, and password-mismatch states use the same response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/users/login": {
      "post": {
        "operationId": "Auth_loginUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthLoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The login credentials are invalid. Missing, locked, passwordless, and password-mismatch states use the same response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "oneOf": [
                            {
                              "$ref": "#/components/schemas/AuthMeCustomerResponseDto"
                            },
                            {
                              "$ref": "#/components/schemas/AuthMeUserResponseDto"
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Homestay Booking Management System API",
    "description": "HTTP contract for the HBMS customer and management applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SuccessEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Thanh cong."
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Response payload. The concrete schema is defined per route."
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "f75cf4ac-6958-4d59-8a4f-183abfeaed15"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "path",
          "timestamp",
          "requestId"
        ]
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "COMMON_VALIDATION_FAILED",
          "COMMON_UNAUTHORIZED",
          "COMMON_FORBIDDEN",
          "COMMON_NOT_FOUND",
          "COMMON_CONFLICT",
          "COMMON_RATE_LIMITED",
          "COMMON_PAYLOAD_TOO_LARGE",
          "COMMON_UNSUPPORTED_MEDIA_TYPE",
          "COMMON_SERVICE_UNAVAILABLE",
          "COMMON_INTERNAL_ERROR",
          "CUSTOMER_INITIAL_PASSWORD_ALREADY_CONFIGURED",
          "CUSTOMER_EMAIL_IN_USE",
          "CUSTOMER_PHONE_IN_USE",
          "CUSTOMER_CURRENT_PASSWORD_INVALID",
          "CUSTOMER_PASSWORD_REUSE_NOT_ALLOWED",
          "AMENITY_NAME_ALREADY_EXISTS",
          "AMENITY_IN_USE",
          "BOOKING_GUEST_CAPACITY_EXCEEDED",
          "BOOKING_CUSTOMER_CONTACT_REQUIRED",
          "BOOKING_ACTIVE_UNPAID_LIMIT_REACHED",
          "BOOKING_HELD_NIGHTS_LIMIT_REACHED",
          "BOOKING_ROOM_NOT_BOOKABLE",
          "BOOKING_CHECKIN_IN_PAST",
          "BOOKING_CHECKIN_TOO_FAR",
          "BOOKING_DATE_RANGE_INVALID",
          "BOOKING_STAY_TOO_LONG",
          "BOOKING_TOTAL_LIMIT_EXCEEDED",
          "BOOKING_ROOM_UNAVAILABLE",
          "BOOKING_CREATE_CONFLICT",
          "BOOKING_CANCELLATION_REASON_REQUIRED",
          "BOOKING_REFUND_PENDING",
          "BOOKING_TRANSITION_NOT_ALLOWED",
          "BOOKING_CONFIRMATION_REQUIRES_PAYMENT",
          "BOOKING_CHECKIN_REQUIRES_PAYMENT",
          "BOOKING_CHECKIN_OUTSIDE_STAY_WINDOW",
          "BOOKING_ROOM_NOT_FOUND",
          "BOOKING_ROOM_NOT_READY",
          "BOOKING_CANCELLATION_ALREADY_PAID",
          "BOOKING_CANCELLATION_NOT_ALLOWED",
          "PAYMENT_REFUND_REJECTED",
          "PAYMENT_REFUND_NOT_ALLOWED",
          "PAYMENT_REFUND_OUTCOME_UNKNOWN",
          "PAYMENT_IDEMPOTENCY_KEY_CONFLICT"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "errorCode": {
            "example": "COMMON_VALIDATION_FAILED",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "example": "Du lieu khong hop le.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "fieldErrors": {
            "type": "object",
            "description": "Validation errors keyed by request field.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "required": [
                  "errorCode"
                ],
                "properties": {
                  "errorCode": {
                    "type": "string",
                    "enum": [
                      "COMMON_VALIDATION_FAILED",
                      "COMMON_UNAUTHORIZED",
                      "COMMON_FORBIDDEN",
                      "COMMON_NOT_FOUND",
                      "COMMON_CONFLICT",
                      "COMMON_RATE_LIMITED",
                      "COMMON_PAYLOAD_TOO_LARGE",
                      "COMMON_UNSUPPORTED_MEDIA_TYPE",
                      "COMMON_SERVICE_UNAVAILABLE",
                      "COMMON_INTERNAL_ERROR",
                      "CUSTOMER_INITIAL_PASSWORD_ALREADY_CONFIGURED",
                      "CUSTOMER_EMAIL_IN_USE",
                      "CUSTOMER_PHONE_IN_USE",
                      "CUSTOMER_CURRENT_PASSWORD_INVALID",
                      "CUSTOMER_PASSWORD_REUSE_NOT_ALLOWED",
                      "AMENITY_NAME_ALREADY_EXISTS",
                      "AMENITY_IN_USE",
                      "BOOKING_GUEST_CAPACITY_EXCEEDED",
                      "BOOKING_CUSTOMER_CONTACT_REQUIRED",
                      "BOOKING_ACTIVE_UNPAID_LIMIT_REACHED",
                      "BOOKING_HELD_NIGHTS_LIMIT_REACHED",
                      "BOOKING_ROOM_NOT_BOOKABLE",
                      "BOOKING_CHECKIN_IN_PAST",
                      "BOOKING_CHECKIN_TOO_FAR",
                      "BOOKING_DATE_RANGE_INVALID",
                      "BOOKING_STAY_TOO_LONG",
                      "BOOKING_TOTAL_LIMIT_EXCEEDED",
                      "BOOKING_ROOM_UNAVAILABLE",
                      "BOOKING_CREATE_CONFLICT",
                      "BOOKING_CANCELLATION_REASON_REQUIRED",
                      "BOOKING_REFUND_PENDING",
                      "BOOKING_TRANSITION_NOT_ALLOWED",
                      "BOOKING_CONFIRMATION_REQUIRES_PAYMENT",
                      "BOOKING_CHECKIN_REQUIRES_PAYMENT",
                      "BOOKING_CHECKIN_OUTSIDE_STAY_WINDOW",
                      "BOOKING_ROOM_NOT_FOUND",
                      "BOOKING_ROOM_NOT_READY",
                      "BOOKING_CANCELLATION_ALREADY_PAID",
                      "BOOKING_CANCELLATION_NOT_ALLOWED",
                      "PAYMENT_REFUND_REJECTED",
                      "PAYMENT_REFUND_NOT_ALLOWED",
                      "PAYMENT_REFUND_OUTCOME_UNKNOWN",
                      "PAYMENT_IDEMPOTENCY_KEY_CONFLICT"
                    ]
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "details": {
            "type": "object",
            "description": "Safe structured context that can help a client recover.",
            "additionalProperties": false,
            "properties": {
              "retryable": {
                "type": "boolean",
                "example": false
              },
              "limit": {
                "type": "integer",
                "example": 3,
                "minimum": 0
              },
              "maxAdvanceDays": {
                "type": "integer",
                "example": 365,
                "minimum": 0
              },
              "maxStayNights": {
                "type": "integer",
                "example": 30,
                "minimum": 0
              }
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "f75cf4ac-6958-4d59-8a4f-183abfeaed15"
          }
        },
        "required": [
          "success",
          "statusCode",
          "errorCode",
          "message",
          "error",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "AuthRegistrationAcceptedDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "example": true,
            "description": "The customer account was created successfully."
          }
        },
        "required": [
          "accepted"
        ]
      },
      "RegisterCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Pham Van Tan",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "tan@example.com",
            "maxLength": 160,
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "0705840355"
          },
          "password": {
            "type": "string",
            "example": "StrongPassword123!",
            "maxLength": 72,
            "minLength": 8,
            "writeOnly": true
          }
        },
        "required": [
          "fullName",
          "phone",
          "password"
        ]
      },
      "AuthCustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "+84912345678"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOCKED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "phone",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Tran Thi B"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+84987654321",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "ADMIN",
            "enum": [
              "STAFF",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOCKED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "phone",
          "role",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthLoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT access token"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ],
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "example": 3600,
            "description": "Token expiration time in seconds"
          },
          "actorType": {
            "type": "string",
            "example": "customer",
            "enum": [
              "customer",
              "user"
            ]
          },
          "customer": {
            "nullable": true,
            "description": "Customer profile if actorType is customer",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthCustomerDto"
              }
            ]
          },
          "user": {
            "nullable": true,
            "description": "User profile if actorType is user",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "actorType"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email or Vietnamese phone number.",
            "example": "tan@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPassword123!",
            "writeOnly": true
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "AuthMeCustomerResponseDto": {
        "type": "object",
        "properties": {
          "actorType": {
            "type": "string",
            "enum": [
              "customer"
            ],
            "example": "customer"
          },
          "customer": {
            "$ref": "#/components/schemas/AuthCustomerDto"
          }
        },
        "required": [
          "actorType",
          "customer"
        ]
      },
      "AuthMeUserResponseDto": {
        "type": "object",
        "properties": {
          "actorType": {
            "type": "string",
            "enum": [
              "user"
            ],
            "example": "user"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "actorType",
          "user"
        ]
      },
...
    }
  }
}